Fix warning: explicit cast for size_t argument
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@31765 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/wp/impexp/xp/ie_exp_HTML_util.cpp
@@ -440,7 +440,7 @@ UT_UTF8String IE_Exp_HTML_MultipartExporter::saveData(const gchar *szDataId,
     
     UT_UTF8String contents;
     encodeDataBase64(szDataId, contents, false);
-    UT_DEBUGMSG(("%lu", contents.length()));
+    UT_DEBUGMSG(("%lu", (long unsigned)contents.length()));
     m_buffer += contents;
     m_buffer += MYEOL;
     m_buffer += MYEOL;
